An MDB (Modeled DataBase) relying on ClickHouse: chMDB

An MDB (Modeled DataBase) relying on ClickHouse: chMDB

Rename tables of a chMDB object

Usage

chMDB(
  tkcon,
  dbTables,
  dbInfo,
  dataModel,
  collectionMembers = NULL,
  check = TRUE,
  n_max = 10,
  verbose = FALSE
)

# S3 method for chMDB
names(x) <- value

# S3 method for chMDB
rename(.data, ...)

# S3 method for chMDB
[(x, i)

# S3 method for chMDB
[[(x, i)

Arguments

tkcon

a chTKCat object

dbTables

a named vector of tables in tkcon$chcon with all(names(dbTables) %in% names(dataModel))

dbInfo

a list with DB information: "name" (only mandatory field), "title", "description", "url", "version", "maintainer".

dataModel

a ReDaMoR::RelDataModel object

collectionMembers

the members of collections as provided to the collection_members<- function (default: NULL ==> no member).

check

logical: if TRUE (default) the data are confronted to the data model

n_max

maximum number of records to read for checks purpose (default: 10). If 0, the data are not checked. See also ReDaMoR::confront_data().

verbose

if TRUE display the data confrontation report

x

a chMDB object

value

new table names

.data

a chMDB object

...

Use new_name = old_name to rename selected tables

i

the index or the name of the tables to take

Value

A chMDB object

See also